+++ packages/ses/test/test-compartment-instance.js
@@ -27,7 +27,14 @@ test('Compartment instance', t => {
   t.deepEqual(Reflect.ownKeys(c), [], 'static properties');
   t.deepEqual(
     Reflect.ownKeys(Object.getPrototypeOf(c)).sort(),
-    ['constructor', 'evaluate', 'name', 'globalThis', 'toString'].sort(),
+    [
+      '__isKnownScopeProxy__',
+      'constructor',
+      'evaluate',
+      'name',
+      'globalThis',
+      'toString',
+    ].sort(),
     'prototype properties',
   );
 });