+++ packages/ses/test/test-module-compartment-instance.js
@@ -29,6 +29,7 @@ test('Compartment instance', t => {
   t.deepEqual(
     Reflect.ownKeys(Object.getPrototypeOf(c)).sort(),
     [
+      '__isKnownScopeProxy__',
       'constructor',
       'evaluate',
       'import',