+++ packages/ses/NEWS.md
@@ -21,7 +21,11 @@ User-visible changes in SES:
   [Tracking issue for getting 3rd party packages more SES friendly (#576)](https://github.com/endojs/endo/issues/576)
   we track the incompatibilities we encounter and progress toward
   getting them fixed.
-
+- Add utility function `Compartment.prototype.__isKnownScopeProxy__(value)` to
+  aid working around scopeProxy leakage. Returns true if `value` is one of the
+  scopeProxies created during calls to this Compartment instances's
+  `Compartment.prototype.evaluate`. See `test-compartment-known-scope-proxy.js`
+  for an example of performing a scopeProxy leak workaround.
 
 ## Release 0.12.3 (1-Mar-2021)